Add votes controller WIP
Add the create vote method in the controller.  This method will need to be changed when I get a better understanding of how the polymorphic association works and where to redirect.  Because there are comment and question show pages the restful pattern is communicating that the vote should redirect to those places, but I can probably infer that we need to be redirecting back to the one page with all the uparrows.