Add partial for vote count
At the moment this is just counting the amount of votes associated with the question.  Really there needs to be a method on the question model that returns the current vote state, as the value is 1 or -1 which will vary depending on the amount of up and down votes.  That number is what should be here instead, but I was at the moment just trying to get ajax partials to render from the controller, which I still can't quite figure out.